Created a FutureWatchList model for users to add media they haven't encountered yet.
Fixed the logic for movies and series for when it is added to a user's watchlist_past

Created a separate function in both view.py files in accounts and review_app apps where users can retrieve movie/tv info using the api.

Changed url pattern 'Movies and TV' to 'movies-tv' in urls.py (category as something more readable in a link) and in other files containing 'Movies and TV'.

TODO:
- look for code that utilizes fetching of data from their api and replace it with getBook, getMovieTV and getGame info.
- make futurewatchlist functional: add buttons conditionally similar to  watchlist_past
- make profile picture in profile.html smaller