Fixed profile's watchlist view. User can now toggle between watchlist and favorites. User can also filter their watchlists.
User can select drop down arrow to pick the category of their watchlist and select if they want to see their past watchlist or future watchlist.

Fixed base_search.html: user will see the movie-tv imdbID instead of the title in the url. Uniformity purposes to make it like the other categories.

Removed the JSON field in the FutureWatchlist model since its redundant. Only saved the id to use it as a lookup value.

Edited get.py for both accounts and review_app to accommodate looking for a movie based on its imdbID like books and games. Integrated this in review_app/views

Changed watchlist_books and watchlist_games to make the cards appear in a user's profile. (html files)

Dropdown for categories: included if-else statement for the button so that the text changes when the category is changed.